Made some organizational modifications. Specifically, created subpackages in the data_access folder to organize the files better.
