+++ src/main/java/callhub/connect/MainController.java
@@ -1,12 +1,9 @@
 package callhub.connect;
 
-import callhub.connect.entities.Message;
-import callhub.connect.data_access.MessageRepository;
+import callhub.connect.data_access.message_data_access.MessageRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.mongodb.repository.config.EnableMongoRepositories;
 import org.springframework.web.bind.annotation.*;
-
-import java.time.LocalDate;
 
 @RestController
 @RequestMapping("/")
